remove types from input file
