+++ src/components/Select/index.tsx
@@ -1,4 +1,5 @@
-import { Character, Monster } from "../../data/input";
+import { Character } from "../../data/character";
+import { Monster } from "../../data/monster";
 import { Build, useBuild } from "../../hooks/useBuild";
 import { capitalize } from "../../utils/format";
 import "./index.css";
